day len git de up len server

+++ wp-content/plugins/woocommerce/includes/admin/settings/class-wc-settings-tax.php
@@ -2,9 +2,7 @@
 /**
  * WooCommerce Tax Settings
  *
- * @author      WooThemes
- * @category    Admin
- * @package     WooCommerce/Admin
+ * @package     WooCommerce\Admin
  * @version     2.1.0
  */
 
@@ -66,6 +64,7 @@ class WC_Settings_Tax extends WC_Settings_Page {
 		$tax_classes = WC_Tax::get_tax_classes();
 
 		foreach ( $tax_classes as $class ) {
+			/* translators: $s tax rate section name */
 			$sections[ sanitize_title( $class ) ] = sprintf( __( '%s rates', 'woocommerce' ), $class );
 		}
 
@@ -82,7 +81,7 @@ class WC_Settings_Tax extends WC_Settings_Page {
 		$settings = array();
 
 		if ( '' === $current_section ) {
-			$settings = include 'views/settings-tax.php';
+			$settings = include __DIR__ . '/views/settings-tax.php';
 		}
 		return apply_filters( 'woocommerce_get_settings_' . $this->id, $settings, $current_section );
 	}
@@ -95,7 +94,7 @@ class WC_Settings_Tax extends WC_Settings_Page {
 
 		$tax_classes = WC_Tax::get_tax_class_slugs();
 
-		if ( 'standard' === $current_section || in_array( $current_section, $tax_classes, true ) ) {
+		if ( 'standard' === $current_section || in_array( $current_section, array_filter( $tax_classes ), true ) ) {
 			$this->output_tax_rates();
 		} else {
 			$settings = $this->get_settings();
@@ -108,12 +107,16 @@ class WC_Settings_Tax extends WC_Settings_Page {
 	 * Save settings.
 	 */
 	public function save() {
+		// phpcs:disable WordPress.Security.NonceVerification.Missing
 		global $current_section;
 
 		if ( ! $current_section ) {
 			$settings = $this->get_settings();
 			WC_Admin_Settings::save_fields( $settings );
 
+			if ( isset( $_POST['woocommerce_tax_classes'] ) ) {
+				$this->save_tax_classes( wp_unslash( $_POST['woocommerce_tax_classes'] ) ); // phpcs:ignore WordPress.Security.ValidatedSanitizedInput.InputNotSanitized
+			}
 		} elseif ( ! empty( $_POST['tax_rate_country'] ) ) {
 			$this->save_tax_rates();
 		}
@@ -123,8 +126,44 @@ class WC_Settings_Tax extends WC_Settings_Page {
 		}
 
 		// Invalidate caches.
-		WC_Cache_Helper::incr_cache_prefix( 'taxes' );
+		WC_Cache_Helper::invalidate_cache_group( 'taxes' );
 		WC_Cache_Helper::get_transient_version( 'shipping', true );
+		// phpcs:enable WordPress.Security.NonceVerification.Missing
+	}
+
+	/**
+	 * Saves tax classes defined in the textarea to the tax class table instead of an option.
+	 *
+	 * @param string $raw_tax_classes Posted value.
+	 * @return null
+	 */
+	public function save_tax_classes( $raw_tax_classes ) {
+		$tax_classes          = array_filter( array_map( 'trim', explode( "\n", $raw_tax_classes ) ) );
+		$existing_tax_classes = WC_Tax::get_tax_classes();
+		$removed              = array_diff( $existing_tax_classes, $tax_classes );
+		$added                = array_diff( $tax_classes, $existing_tax_classes );
+
+		foreach ( $removed as $name ) {
+			WC_Tax::delete_tax_class_by( 'name', $name );
+		}
+
+		foreach ( $added as $name ) {
+			$tax_class = WC_Tax::create_tax_class( $name );
+
+			// Display any error that could be triggered while creating tax classes.
+			if ( is_wp_error( $tax_class ) ) {
+				WC_Admin_Settings::add_error(
+					sprintf(
+						/* translators: 1: tax class name 2: error message */
+						esc_html__( 'Additional tax class "%1$s" couldn\'t be saved. %2$s.', 'woocommerce' ),
+						esc_html( $name ),
+						$tax_class->get_error_message()
+					)
+				);
+			}
+		}
+
+		return null;
 	}
 
 	/**
@@ -159,17 +198,21 @@ class WC_Settings_Tax extends WC_Settings_Page {
 					'page'    => 'wc-settings',
 					'tab'     => 'tax',
 					'section' => $current_section,
-				), 'admin.php'
+				),
+				'admin.php'
 			)
 		);
 
 		// Localize and enqueue our js.
 		wp_localize_script(
-			'wc-settings-tax', 'htmlSettingsTaxLocalizeScript', array(
+			'wc-settings-tax',
+			'htmlSettingsTaxLocalizeScript',
+			array(
 				'current_class' => $current_class,
 				'wc_tax_nonce'  => wp_create_nonce( 'wc_tax_nonce-class:' . $current_class ),
 				'base_url'      => $base_url,
 				'rates'         => array_values( WC_Tax::get_rates_for_tax_class( $current_class ) ),
+				// phpcs:ignore WordPress.Security.NonceVerification.Recommended
 				'page'          => ! empty( $_GET['p'] ) ? absint( $_GET['p'] ) : 1,
 				'limit'         => 100,
 				'countries'     => $countries,
@@ -206,7 +249,7 @@ class WC_Settings_Tax extends WC_Settings_Page {
 		);
 		wp_enqueue_script( 'wc-settings-tax' );
 
-		include 'views/html-settings-tax.php';
+		include __DIR__ . '/views/html-settings-tax.php';
 	}
 
 	/**
@@ -247,6 +290,7 @@ class WC_Settings_Tax extends WC_Settings_Page {
 			'tax_rate_priority',
 		);
 
+		// phpcs:disable WordPress.Security.NonceVerification.Missing
 		foreach ( $tax_rate_keys as $tax_rate_key ) {
 			if ( isset( $_POST[ $tax_rate_key ], $_POST[ $tax_rate_key ][ $key ] ) ) {
 				$tax_rate[ $tax_rate_key ] = wc_clean( wp_unslash( $_POST[ $tax_rate_key ][ $key ] ) );
@@ -257,6 +301,7 @@ class WC_Settings_Tax extends WC_Settings_Page {
 		$tax_rate['tax_rate_shipping'] = isset( $_POST['tax_rate_shipping'][ $key ] ) ? 1 : 0;
 		$tax_rate['tax_rate_order']    = $order;
 		$tax_rate['tax_rate_class']    = $class;
+		// phpcs:enable WordPress.Security.NonceVerification.Missing
 
 		return $tax_rate;
 	}
@@ -267,7 +312,8 @@ class WC_Settings_Tax extends WC_Settings_Page {
 	public function save_tax_rates() {
 		global $wpdb;
 
-		$current_class    = sanitize_title( $this->get_current_tax_class() );
+		$current_class = sanitize_title( $this->get_current_tax_class() );
+		// phpcs:ignore WordPress.Security.ValidatedSanitizedInput.InputNotValidated, WordPress.Security.NonceVerification.Missing
 		$posted_countries = wc_clean( wp_unslash( $_POST['tax_rate_country'] ) );
 
 		// get the tax rate id of the first submited row.
@@ -279,13 +325,14 @@ class WC_Settings_Tax extends WC_Settings_Page {
 		$index = isset( $tax_rate_order ) ? $tax_rate_order : 0;
 
 		// Loop posted fields.
+		// phpcs:disable WordPress.Security.NonceVerification.Missing
 		foreach ( $posted_countries as $key => $value ) {
 			$mode     = ( 0 === strpos( $key, 'new-' ) ) ? 'insert' : 'update';
 			$tax_rate = $this->get_posted_tax_rate( $key, $index ++, $current_class );
 
 			if ( 'insert' === $mode ) {
 				$tax_rate_id = WC_Tax::_insert_tax_rate( $tax_rate );
-			} elseif ( 1 === absint( $_POST['remove_tax_rate'][ $key ] ) ) {
+			} elseif ( isset( $_POST['remove_tax_rate'][ $key ] ) && 1 === absint( $_POST['remove_tax_rate'][ $key ] ) ) {
 				$tax_rate_id = absint( $key );
 				WC_Tax::_delete_tax_rate( $tax_rate_id );
 				continue;
@@ -301,6 +348,7 @@ class WC_Settings_Tax extends WC_Settings_Page {
 				WC_Tax::_update_tax_rate_cities( $tax_rate_id, wc_clean( wp_unslash( $_POST['tax_rate_city'][ $key ] ) ) );
 			}
 		}
+		// phpcs:enable WordPress.Security.NonceVerification.Missing
 	}
 }
